% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{compare_proxy}
\alias{compare_proxy}
\title{Proxy for waldo comparison}
\usage{
compare_proxy(x)
}
\arguments{
\item{x}{An object.}
}
\description{
Use this generic to override waldo's default comparison if you need to
override the defaults (typically because your object contains an external
pointer).

waldo comes with methods for two common cases:
\itemize{
\item data.table: the \code{.internal.selfref} attribute is set to \code{NULL}. This is
an external pointer that is used for performance optimisation, and
doesn't affect the data.
\item \code{xml2::xml_node}: the underlying XML data is stored in memory in C,
behind an external pointer, so the we best can do is to convert the
object to a string.
}
}
