% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{diff}
\alias{diff}
\title{Show the difference between two objects}
\usage{
diff(old, new, ...)
}
\arguments{
\item{old, new}{Pair of objects to compare. The first argument is taken
to be the "old" value, and the second is the "new". This is effectively
the opposite to \code{\link[=compare]{compare()}}.}

\item{...}{Other arguments passed on to \code{\link[=compare]{compare()}}}
}
\description{
Show the difference between two objects
}
\examples{
x1 <- x2 <- list(x = 1, y = 2:10, z = letters)
x2$z <- letters[-13]
diff(x1, x2)
}
