% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waldo-package.R
\docType{package}
\name{waldo-package}
\alias{waldo}
\alias{waldo-package}
\title{waldo: Find Differences Between R Objects}
\description{
Compare complex R objects and reveal the key differences. Designed particularly for use in testing packages where being able to quickly isolate key differences makes understanding test failures much easier.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://waldo.r-lib.org}
  \item \url{https://github.com/r-lib/waldo}
  \item Report bugs at \url{https://github.com/r-lib/waldo/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
